Merge|GameEngine: Implement decor rendering into pipeline

+++ map-generator/DecorHandling/DecorPosition.cs
@@ -5,7 +5,7 @@ namespace map_generator.DecorHandling;
 /**A data class for storing a Decor and its position/rotation*/
 public record DecorPosition(
     [property: JsonPropertyName("name"), JsonConverter(typeof(DecorJsonConverter))] Decor Decor,
-    [property: JsonPropertyName("x")] double XPos,
-    [property: JsonPropertyName("y")] double YPos,
-    double Rotation
+    [property: JsonPropertyName("x")] float XPos,
+    [property: JsonPropertyName("y")] float YPos,
+    float Rotation
 );
